requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.3
django-allauth==64.0.0
django-anymail==11.0.1
django-appconf==1.0.6
django-celery-email==3.0.0
django-cors-headers==4.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
emoji==2.14.1
empire-chain==0.4.3
eval_type_backport==0.2.2
executing==2.1.0
faiss-cpu==1.10.0
Faker==25.8.0
fastapi==0.115.7
fastjsonschema==2.20.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
groq==0.15.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
grpcio-tools==1.70.0
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
h5py==3.14.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.7
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.10.0
kiwisolver==1.4.7
kombu==5.4.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.32
libclang==18.1.1
lightgbm==4.6.0
lxml==5.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
Mean_Squared_Error==0.4.0
mistune==3.0.2
ml_dtypes==0.5.1
mlxtend==0.23.4
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
namex==0.1.0
narwhals==1.24.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
olefile==0.47
openai==1.75.0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.16
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.3.0
platformdirs==4.3.6
portalocker==2.10.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.25.5
pyparsing==3.2.0
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-json-logger==2.0.7
python-magic==0.4.27
python-oxmsg==0.0.2
python-whois==0.9.5
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.2
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==75.2.0
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.17.2
unstructured-client==0.33.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
vectorizer==0.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
word2number==1.1
wordcloud==1.9.4
wrapt==1.17.2
xgboost==3.0.0
yarl==1.20.0
zstandard==0.23.0
